// src/app/models/myAds.ts
export class MyAds {
    active: string;
    ad_status: string;
    area: string;
    area_type: string;
    category: string;
    category_id: string;
    category_title: string;
    commercial: string;
    commercial_price: string;
    contact_number: string;
    created: string;
    creation_date: string;
    currency: string;
    days_published: string;
    description: string;
    directory: string;
    display_description: string;
    display_title: string;
    email: string;
    expiration_date: string;
    expiration_date_diff: string;
    has_image: string;
    hide_email: string;
    housing_available_bedrooms: string;
    housing_available_period: string;
    housing_available_price: string;
    item_price: string;
    last_modified: string;
    make_model: string;
    new_region: string;
    odometer: string;
    parent_category_id: string;
    payment: string;
    personals_age: string;
    picture1: string;
    picture2: string;
    picture3: string;
    picture4: string;
    picture5: string;
    picture6: string;
    picture7: string;
    picture8: string;
    posting_id: string;
    premium: string;
    premium_days_published: string;
    preview_desc: string;
    published_date: string;
    published_date_epoch: string;
    region: string;
    region_code: string;
    region_id: string;
    region_name: string;
    show_adsense: string;
    sms_telephone_number: string;
    title: string;
    transmission: string;
    unique_id: string;
    year: string;

    constructor(){
      this.active = "";
      this.ad_status = "";
      this.area = "";
      this.area_type = "";
      this.category = "";
      this.category_id = "";
      this.category_title = "";
      this.commercial = "";
      this.commercial_price = "";
      this.contact_number = "";
      this.created = "";
      this.creation_date = "";
      this.currency = "";
      this.days_published = "";
      this.description = "";
      this.directory = "";
      this.display_description = "";
      this.display_title = "";
      this.email = "";
      this.expiration_date = "";
      this.expiration_date_diff = "";
      this.has_image = "";
      this.hide_email = "";
      this.housing_available_bedrooms = "";
      this.housing_available_period = "";
      this.housing_available_price = "";
      this.item_price = "";
      this.last_modified = "";
      this.make_model = "";
      this.new_region = "";
      this.odometer = "";
      this.parent_category_id = "";
      this.payment = "";
      this.personals_age = "";
      this.picture1 = "";
      this.picture2 = "";
      this.picture3 = "";
      this.picture4 = "";
      this.picture5 = "";
      this.picture6 = "";
      this.picture7 = "";
      this.picture8 = "";
      this.posting_id = "";
      this.premium = "";
      this.premium_days_published = "";
      this.preview_desc = "";
      this.published_date = "";
      this.published_date_epoch = "";
      this.region = "";
      this.region_code = "";
      this.region_id = "";
      this.region_name = "";
      this.show_adsense = "";
      this.sms_telephone_number = "";
      this.title = "";
      this.transmission = "";
      this.unique_id = "";
      this.year = "";
      
    }
  }